Every 'mix test' problem
Everytime I do 'mix test', the tests are run in different order or so I see in the terminal. Doing async false helps because they are run sequential but the start order is still different and async false takes longer because each tests needs to finish before the other test can run.
Sometimes I see CoinDataRetriever test first and it fails because its not waiting and elem(1) gets nil as argument because Processmanager doesnt have pais yet because the retrieve coin pairs hasn't finished yet etc

Sleeping for 1 second and everything is OK